refactor: extract runtime shortcode theme concern
